Initial commit
I'm not even close being done here
Let's start taking out the obvious: WSL and Linux aren't the same thing, not even close. While WSL acts pretty similar to linux, some staff just doesn't work, and while usually it's not that big of a deal, in this case it is, since it messes with psutil and forces me to do stuff in a non-efficient way. Which is a problem, since I need this to be working under the WSL. That also means that until I have a working version, moving to a real linux os to test is not do-able, since it would be a huge loss of time.
File names have to be fixes, and I really need to find a way to obtain the process target architecture. Logs are to be done, too, together with proper argument processing, popen command handling, etc. After this, it could be useful to propose as an alternative json to xlsx - excel worksheets are useful in certain scenarios, but I don't think I want this to work only on that. I may want to pass the data I collect to something else.
Setup is also a TBD, together with local install scripts for both bash and cmd.