Almost functional prototype
The argument parser is up and pretty much done. I had some troubles parsing the command line, so I had to resort by treating commands differently depending if I'm running on Windows or somewhere else. The logger is also there, and there's a raw print on stdout.
The main thing that doesn't really check out is the memory usage statistic, it looks a bit too off on Windows, especially for commands that take only a couple of seconds to run.
Now I only need total time and effective file reporting