Trimmed mean is available
Turns out that outliers have to be taken in consideration whatever I do to reduce the errors. This will likely double the memory consumption at runtime. If that turns out to be a problem I'll introduce a way to split the work in smaller section with periodical dumps on file. I still haven't implemented a way to "unify" tests, tho, and I feel like part of that functionality might be useful here